Update Flex Panel
<h4>Flex Panel</h4>
The Window of this website comprises of 5 panels of that expand and shrink on been clicked. It also brings in transition of the top and bottom words into the window.

<h5>How it looks?</h5>
Here's the answer to it!!
![flex panel](https://user-images.githubusercontent.com/39344406/42477723-3b9cae66-83a0-11e8-9b40-9bd7bd6c4b1c.png)

And on being clicked, a click event gets fired that is fetched by JS and modified on CSS3. The panel whose click event is fired grows!


<h5>What and How?</h5>
An HTMl file executes this(if you want JS and CSS inline, this is all you need; or check the note below), and to create it any editor(top geeky or plain old default) can help.
I am using Visual Code as it shows all directories on the panel.
<i>Check out FlexPanel.html</i> 

<h6>NOTE:</h6>If you wish to separate CSS from HTML, use <link> tag; and for separate JS file use <script> tag.

Secondly, Java script code that fetches the click event fired boils down to this-

<b>const panels = document.querySelectorAll('.panel');
    function toggleOpen(){
        this.classList.toggle('open');
    }
    function toggleActive(e){
        if(e.propertyName.includes('flex')){
        this.classList.toggle('open-active');
    }}
    panels.forEach(panel=> panel.addEventListener('click',toggleOpen));
    panels.forEach(panel=> panel.addEventListener('transitionend',toggleActive));
</b>

To save the reader from gibber, here's the break down-
<b>panels</b> is a constant that saves value from HTML file's <div> with class called "panel". In the last two lines this constant is used to run a loop and add events with associated method. The two methods <b>toggleOpen()</b> and <b>toggleActive(e)</b> would toggle(cant be simpler than this) the class list with new attribute.

Thirdly, the CSS does rest of the magic. The codes are-
The words that appear from the top and bottom are transformed on Y-Axis
<b>
.panel>*:first-child{transform: translateY(-1000%);}
.panel.open-active>*:first-child{transform: translateY(0);}
.panel>*:last-child{transform: translateY(1000%);}
.panel.open-active>*:last-child{transform: translateY(0);}
</b>
and the panel grows with this transition-
<b>
transition: 
        font-size 0.7s cubic-bezier(0.13, 0.64, 0.43, 0.65),
        flex-grow 0.7s cubic-bezier(0.13, 0.64, 0.43, 0.65),
        background 0.2s;
along with
.panel.open{
    flex:5;
    font-size: 20px;
}
</b>


<h5>Acknowledgments</h5>
The brains behind are detailed tutorial by JavaScript30.com, Stack Overflow and Fun Fun Function.